Fix all clippy warnings across lib and test code
- items_after_test_module: move decode_huffman before #[cfg(test)] in decode.rs
- approx_constant: replace PI/E-approximate f32 literals with non-constant values
- needless_range_loop: convert index loops to iterator patterns in tests
- unnecessary_cast: remove redundant i32 casts
- let_and_return: return expression directly
- vec_init_then_push: use vec![..] literal
- useless_vec: use array literal instead of vec!
- unused_variable: prefix unused binding with underscore

+++ tests/decode_into.rs
@@ -114,7 +114,7 @@ test_decode_into_matches!(
     i32,
     0i32,
     (0..32 * 32)
-        .map(|i| i as i32 * 100 - 50000)
+        .map(|i| i * 100 - 50000)
         .collect::<Vec<i32>>(),
     32,
     32,
